% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAplot.R
\name{BAplot}
\alias{BAplot}
\title{BAplot}
\usage{
BAplot(df)
}
\arguments{
\item{df}{A data frame returned by the BAtable() function}
}
\value{
A Bland Altman plot
}
\description{
Draws a Bland Altman plot using the output of the BAtable() function
}
