% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAtable.R
\name{BAtable}
\alias{BAtable}
\title{BAtable}
\usage{
BAtable(x, y)
}
\arguments{
\item{x}{A vector of linear assessment scores}

\item{y}{A vector of CAT assessment scores}
}
\value{
A data frame
}
\description{
Creates a data frame of linear assessment factor scores, CAT factor scores, and the mean and difference of paired scores. Used for calculating limits of agreement and creating Bland Altman plots.
}
