% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLimits.R
\name{getLimits}
\alias{getLimits}
\title{getLimits}
\usage{
getLimits(BAtable)
}
\arguments{
\item{BAtable}{A dataframe produced by the BAtable() function}
}
\value{
Lower and upper 95% limits of agreement (LoA)
}
\description{
Calculates 95% limits of agreement
}
