% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRasch.R
\name{getRasch}
\alias{getRasch}
\title{getRasch}
\usage{
getRasch(list, df)
}
\arguments{
\item{list}{A value returned from the scaleInfo() function}

\item{df}{A dataframe containing calibration item responses where each row represents a respondent and each column represents an item}
}
\value{
Returns Rasch model fit statistics and item parameters
}
\description{
Provides Rasch model item parameters
}
