% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{impute}
\usage{
impute(df)
}
\arguments{
\item{df}{A dataframe containing item responses with missing data labelled as NA}
}
\value{
Returns the imputed dataset
}
\description{
Imputes missing item responses using multiple imputation with chained equations
}
