% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAssessment.R
\name{linearAssessment}
\alias{linearAssessment}
\title{linearAssessment}
\usage{
linearAssessment(scale, responses)
}
\arguments{
\item{scale}{Specifies which scale to use. Options are: "Face", "Jaw", "Teeth", "School", "Psych", "SDistress" (for speech distress), "SFunction" (for speech functin), or "Social"}

\item{responses}{A matrix of responses, where each column represents an item from the full-length scale (ordered 1, 2, 3...) and each row represents a respondent}
}
\value{
Returns a list of results from the mirtCAT::mirtCAT() function
}
\description{
Simulates a linear assessment for a specified CLEFT-Q scale and set of responses
}
