% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noOutliers.R
\name{noOutliers}
\alias{noOutliers}
\title{noOutliers}
\usage{
noOutliers(matrix)
}
\arguments{
\item{matrix}{A matrix of response options, where each column represents an item and each row represents a respondent}
}
\value{
Returns a matrix with outliers removed
}
\description{
Remove outliers from a validation dataset by Mahalanobis distance
}
