% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBA.R
\name{plotBA}
\alias{plotBA}
\title{plotBA}
\usage{
plotBA(scale, responses, stopping_rule)
}
\arguments{
\item{scale}{Specifies which scale to use. Options are: "Face", "Jaw", "Teeth", "School", "Psych", "SDistress" (for speech distress), "SFunction" (for speech functin), or "Social"}

\item{responses}{A matrix of responses, where each column represents an item from the full-length scale (ordered 1, 2, 3...) and each row represents a respondent}

\item{stopping_rule}{A fixed-length stopping rule (number of items as an integer value)}
}
\value{
A Bland Altman plot
}
\description{
Draws a Bland Altamn plot comparing transformed (0-100) CAT scores and linear assessments, for a given scale at a given fixed length stopping rule
}
