% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescore.R
\name{rescore}
\alias{rescore}
\title{rescore}
\usage{
rescore(df)
}
\arguments{
\item{df}{A dataframe containing item responses (1,2,3 and 4) to the field test version of CLEFT-Q Speech Distress and Speech Function scales}
}
\value{
The rescored item responses
}
\description{
Rescore item responses in the Speech Distress and Speech Function scales (collapses response options 2 and 3, then reverses scoring)
}
