% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleInfo.R
\name{scaleInfo}
\alias{scaleInfo}
\title{scaleInfo}
\usage{
scaleInfo(chr)
}
\arguments{
\item{chr}{The name of a scale included in the 2017 ICHOM Standard Set for cleft lip and/or palate. Possible arguments include: "Face", "Jaw", "Teeth", "School", "Psych", "SDistress", "SFunction" or "Social"}
}
\value{
A list containing scale name, validation response file name, calibration dataset columns corresponding to item responses, minimum factor score (EAP), maximum factor score (EAP) and total number of items.
}
\description{
Gets all scale information required for Monte Carlo CAT simulations
}
