% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDistribution.R
\name{scoreDistribution}
\alias{scoreDistribution}
\title{scoreDistribution}
\usage{
scoreDistribution(scale, result)
}
\arguments{
\item{scale}{Specifies which scale to use. Options are: "Face", "Jaw", "Teeth", "School", "Psych", "SDistress" (for speech distress), "SFunction" (for speech functin), or "Social"}

\item{result}{A list resturned by the linearAssessmet() function}
}
\value{
A kernal density plot of factor scores
}
\description{
Plots factor scores as a kernal density plot
}
