% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOutliers.R
\name{showOutliers}
\alias{showOutliers}
\title{showOutliers}
\usage{
showOutliers(matrix)
}
\arguments{
\item{matrix}{A matrix of item responses, where each row represents a respondent and each column represents an item}
}
\value{
Returns sample size, number of outliers, X2 test statistic and degrees of freedom
}
\description{
Counts outliers by Mahalanobis distance in a sample of item responses
}
