% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCATlogits.R
\name{simCATlogits}
\alias{simCATlogits}
\title{simCATlogits}
\usage{
simCATlogits(scale, responses, linear_result)
}
\arguments{
\item{scale}{Specifies which scale to use. Options are: "Face", "Jaw", "Teeth", "School", "Psych", "SDistress" (for speech distress), "SFunction" (for speech functin), or "Social"}

\item{responses}{A matrix of responses, where each column represents an item from the full-length scale (ordered 1, 2, 3...) and each row represents a respondent}

\item{linear_result}{A list returned by the linearAssessment() function}
}
\value{
A data frame of results
}
\description{
Performs a series of Monte Carlo simulations, using CATs developed from the Rasch models in the project folder. All possible fixed length stopping rules are tested for each scale. Expected a posteriori factor scores are generated and presented as logits. Item selection is by minimum expected posterior variance. CAT scores are compared to linear assessment scores through Pearson's correlation coefficient, RMSE and 95% limits of agreement. Standard error of measurement values for each score are also presented.
}
