% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCATtransformed.R
\name{simCATtransformed}
\alias{simCATtransformed}
\title{simCATtransformed}
\usage{
simCATtransformed(scale, responses, linear_result)
}
\arguments{
\item{scale}{Specifies which scale to use. Options are: "Face", "Jaw", "Teeth", "School", "Psych", "SDistress" (for speech distress), "SFunction" (for speech functin), or "Social"}

\item{responses}{A matrix of responses, where each column represents an item from the full-length scale (ordered 1, 2, 3...) and each row represents a respondent}

\item{linear_result}{A list returned by the linearAssessment() function}
}
\value{
A data frame of results
}
\description{
Similar to simCATlogits, but returns RMSE and 95% limits of agreement in transformed 0-100 CLEFT-Q units
}
