% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaBySEM.R
\name{thetaBySEM}
\alias{thetaBySEM}
\title{thetaBySEM}
\usage{
thetaBySEM(scale, result)
}
\arguments{
\item{scale}{Specifies which scale to use. Options are: "Face", "Jaw", "Teeth", "School", "Psych", "SDistress" (for speech distress), "SFunction" (for speech functin), or "Social"}

\item{result}{A list resturned by the linearAssessmet() function}
}
\value{
A scatterplot of SEM by factor score
}
\description{
Plots standard error of measurement (SEM) against factor score
}
